+++ src/Builder/DE/BuilderInterface.php
@@ -90,4 +90,8 @@ interface BuilderInterface
      * Grupo E8.1.1 Campos que describen los descuentos, anticipos y valor total por ítem.
      */
     public function setGroupE811($data);
+    /**
+     * Grupo E8.2 Campos que describen el IVA de la operación por ítem.
+     */
+    public function setGroupE82($data);
 }